requirements.txt
requests
beautifulsoup4
googlesearch-python
thefuzz
python-Levenshtein
ollama
